Moved factories to tests directory
