+++ tests/App/Factory/ProjectFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PhpSentinel\BugCatcher\Factory;
+namespace App\Factory;
 
 use PhpSentinel\BugCatcher\Entity\Project;
 use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
