+++ tests/Integration/Twig/Detail/HistoryListTest.php
@@ -7,7 +7,7 @@
  */
 namespace PhpSentinel\BugCatcher\Tests\Integration\Twig\Detail;
 
-use PhpSentinel\BugCatcher\Factory\RecordLogFactory;
+use App\Factory\RecordLogFactory;
 use PhpSentinel\BugCatcher\Tests\App\KernelTestCase;
 use PhpSentinel\BugCatcher\Twig\Components\Detail\HistoryList;
 use Symfony\UX\TwigComponent\Test\InteractsWithTwigComponents;
