+++ tests/Integration/Twig/Detail/StackTraceTest.php
@@ -8,9 +8,9 @@
 namespace PhpSentinel\BugCatcher\Tests\Integration\Twig\Detail;
 
 
+use App\Factory\RecordLogTraceFactory;
 use Exception;
 use Kregel\ExceptionProbe\Codeframe;
-use PhpSentinel\BugCatcher\Factory\RecordLogTraceFactory;
 use PhpSentinel\BugCatcher\Tests\App\KernelTestCase;
 use PhpSentinel\BugCatcher\Twig\Components\Detail\StackTrace;
 use Symfony\UX\TwigComponent\Test\InteractsWithTwigComponents;
