+++ tests/Integration/Twig/Detail/TitleTest.php
@@ -8,7 +8,7 @@
 
 namespace PhpSentinel\BugCatcher\Tests\Integration\Twig\Detail;
 
-use PhpSentinel\BugCatcher\Factory\RecordLogTraceFactory;
+use App\Factory\RecordLogTraceFactory;
 use PhpSentinel\BugCatcher\Tests\App\KernelTestCase;
 use Symfony\UX\TwigComponent\Test\InteractsWithTwigComponents;
 use Zenstruck\Foundry\Test\Factories;
