+++ tests/Integration/Twig/LogList/RecordLogTest.php
@@ -7,9 +7,8 @@
  */
 namespace PhpSentinel\BugCatcher\Tests\Integration\Twig\LogList;
 
+use App\Factory\RecordLogFactory;
 use DateTime;
-use PhpSentinel\BugCatcher\Factory\RecordLogFactory;
-use PhpSentinel\BugCatcher\Factory\RecordLogTraceFactory;
 use PhpSentinel\BugCatcher\Tests\App\KernelTestCase;
 use PhpSentinel\BugCatcher\Twig\Components\LogList\RecordLog;
 use Symfony\UX\TwigComponent\Test\InteractsWithTwigComponents;
